Rename hooks/checkoffs.js to hoos/pages.js
